Example Images For Extension Working and Illustration.
In this example, we check the index of difficulty between the links of "Pull Requests" And "Explore" on github.

We take 420,35 (x,y) point on Pull requests as first point.
We take 649,35 (xy) point on Explore..

We can say that when user visited github site, his cursor was at pull requests link and point is 420,35. He want to click on explore link. So he move from 420,35 (Pull links) to 649,35 (Explore).

The width of Explore link is, either you put directly or you can add the starting and ending point of explore link on X axis. Explore link is from 629 to 675. So (675-629 = 46 Pixels).

After giving values you will get index of difficulty.
The index of difficulty between Git Hub's Pull Requests and Explore link is approximately 3.31